refactor: move extern_type conversions to validation module
